rename Aws3Config(#156) 파일 위치이동 및 이름 수정
